+++ R/ordering.R
@@ -21,7 +21,7 @@ setMethod("ordering", "CrunchDataset", function(x) ordering(allVariables(x)))
 #' @rdname ordering
 #' @export
 setMethod("ordering<-", "CrunchDataset", function(x, value) {
-    if (is.unforcedVarCat(x@variables)) {
+    if (is.unforcedVariableCatalog(x@variables)) {
         x@variables <- getDatasetVariables(x)
     }
     ordering(x@variables) <- value