+++ tests/testthat/test-as-data-frame.R
@@ -132,7 +132,7 @@ with_mock_crunch({
     })
 
     test_that("as.data.frame when a variable has an apostrophe in its alias", {
-        t2 <- forceVarCat(ds)
+        t2 <- forceVariableCatalog(ds)
 
         t2@variables@index[[2]]$alias <- "Quote 'unquote' alias"
         expect_is(as.data.frame(t2), "CrunchDataFrame")
@@ -323,7 +323,7 @@ with_test_authentication({
 
     v2 <- ds$v2
     # Force variable catalog so that it'll get stale
-    ds <- forceVarCat(ds)
+    ds <- forceVariableCatalog(ds)
     with_consent(delete(v2))
     test_that("CrunchDataFrame lazily fetches columns", {
         expect_true("v2" %in% names(ds)) ## ds is stale
@@ -332,7 +332,7 @@ with_test_authentication({
         expect_error(as.data.frame(ds, force = TRUE))
     })
 
-    ds <- forceVarCat(ds)
+    ds <- forceVariableCatalog(ds)
     uncached({
         with_mock(`crunch::.crunchPageSize` = function(x) 5L, {
             with(temp.option(httpcache.log = ""), {