+++ tests/testthat/test-cube-errors.R
@@ -180,7 +180,7 @@ with_test_authentication({
     })
 
     test_that("Reserved function names cannot be variable aliases", {
-        dsb <- forceVarCat(ds)
+        dsb <- forceVariableCatalog(ds)
         al <- aliases(dsb@variables)
         dsb@variables@index[[which(al == "v1")]]$alias <- "mean"
         dsb@variables@index[[which(al == "v2")]]$alias <- "sd"