+++ tests/testthat/test-dataset-entity.R
@@ -291,7 +291,7 @@ with_mock_crunch({
     })
 
     test_that("Dataset ncol doesn't make any requests if variables have been forced", {
-        ds <- forceVarCat(ds)
+        ds <- forceVariableCatalog(ds)
         with(temp.options(httpcache.log = ""), {
             logs <- capture.output(nc <- ncol(ds))
         })